\name{package_web}
\alias{package_web}
\title{A demo for browsing R packages.}
\description{
This package provides a demo for opening packages' webpage and reference manual in  your browser.
}
\usage{
package_web(package_name)
}
\arguments{
  \item{package_name}{name of package ypu want to browser}
}
\details{
nothing
}
\value{
opening webpage in your browser
}
\references{
https://cran.r-project.org/web/packages/
}
\author{
Summer Zhang
}
\note{
This is a function which I making for practicing ,  pls forgive my poor coding skill and english,thx.
}
\seealso{
nothing
}
\examples{
package_web("ggplot2")
}
\keyword{ ~package }
\keyword{ ~web }
